consolidate Drawer components
